tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-name_none.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - 'none' keyword in animation-name</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-name-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="If the animation-name is 'none' then there will
   be no animation">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes grow-width {
         from {
            width: 100px;
         }
         to {
            width: 200px;
         }
      }
      @keyframes grow-height {
         from {
            height: 100px;
         }
         to {
            height: 200px;
         }
      }
      @keyframes none {
         from {
            background-color: red;
         }
         to {
            background-color: red;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
      }
      .test {
         animation-fill-mode: forwards, forwards, forwards;
         animation-duration: 1s, 1s, 1s;
         animation-name: grow-width, none, grow-height;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var growStartEvents = 0;
         var badEventFired = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function(evt) {
            if (evt.animationName.match(/^grow-(width|height)$/)) {
               growStartEvents++;
            } else {
               badEventFired = true;
            }
         });
         div.addEventListener('animationend', function(evt) {
            clearTimeout(timeout);
            if (!evt.animationName.match(/^grow-(width|height)$/)) {
               badEventFired = true;
            }
            var passed = (growStartEvents == 2 && !badEventFired);
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should grow in size, while remaining square and blue.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
